! BREAKING CHAGE: Dog table move to users
