+++ hospital/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
-
+from .models import Hospital, Reservation
 # Register your models here.
+admin.site.register(Hospital)
+admin.site.register(Reservation)
