+++ hospital/serializers.py
@@ -35,22 +35,6 @@ class HospitalSerializer(serializers.ModelSerializer):
             'updated_at',
         ]
 
-class DogSerializer(serializers.ModelSerializer):
-    user = serializers.SerializerMethodField(read_only=True)
-    
-    def get_user(self, instance):
-        return instance.user.username
-
-    class Meta:
-        model = Dog
-        fields = '__all__'
-        read_only_fields = [
-            'id',
-            'user',
-            'created_at',
-            'updated_at',
-        ]
-
 class ReservationSerializer(serializers.ModelSerializer):
     hospital = serializers.CharField(source='hospital.name', read_only=True)
     user = serializers.CharField(source='user.username', read_only=True)
