+++ hospital/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, include
 from rest_framework import routers
-from .views import HospitalViewSet, DogViewSet, ReservationUserViewSet, ReservationViewSet, HospitalReservationViewSet
+from .views import HospitalViewSet, ReservationUserViewSet, ReservationViewSet, HospitalReservationViewSet
 
 from django.conf import settings
 from django.conf.urls.static import static
@@ -9,9 +9,6 @@ app_name = "hospital"
 
 default_router = routers.SimpleRouter(trailing_slash=False)
 default_router.register('home', HospitalViewSet, basename='home')
-
-dog_router = routers.SimpleRouter(trailing_slash=False)
-dog_router.register('dog', DogViewSet, basename='dog')
 
 reservation_check_router = routers.SimpleRouter(trailing_slash=False)
 reservation_check_router.register('check', ReservationUserViewSet, basename='check')
@@ -24,7 +21,6 @@ hospital_reservation_router.register('reservation', HospitalReservationViewSet, 
 
 urlpatterns = [
     path('', include(default_router.urls)),
-    path('', include(dog_router.urls)),
     path('reservation/', include(reservation_check_router.urls)),
     path('', include(reservation_router.urls)),
     path('home/<int:hospital_id>/', include(hospital_reservation_router.urls)),
