+++ users/views.py
@@ -1,0 +1,20 @@
+from .models import *
+from rest_framework import viewsets, mixins
+from .serializers import DogSerializer
+from rest_framework.response import Response
+
+class DogViewSet(viewsets.ModelViewSet):
+    queryset = Dog.objects.all()
+    serializer_class = DogSerializer
+
+    def list(self, request, user_id=None):
+        user = self.request.user
+        queryset = self.filter_queryset(self.get_queryset().filter(user=user))
+        serializer = self.get_serializer(queryset, many=True)
+        return Response(serializer.data)
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+    def perform_update(self, serializer):
+        serializer.save(user=self.request.user)
